requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==1.10.0
appdirs==1.4.4
argh==0.26.2
astor==0.8.1
async-timeout==3.0.1
atomicwrites==1.4.0
backports.weakref==1.0.post1
base58==2.1.0
billiard==3.6.3.0
bkcharts==0.2
black==19.10b0
blessings==1.7
blinker==1.4
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.2.1
catboost==0.25
category-encoders==2.2.2
certifi==2020.12.5
clyent==1.2.2
coala-utils==0.5.1
colorful==0.5.4
comtypes==1.1.8
conda==4.9.2
conda-build==3.21.4
conda-verify==3.4.2
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.11.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
Flask-MySQLdb==0.2.0
future==0.18.2
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.5
GitPython==3.1.13
gmpy2==2.0.8
google-api-core==1.26.1
google-auth==1.27.1
googleapis-common-protos==1.53.0
gpustat==0.6.0
graphviz==0.16
grpcio==1.36.1
gunicorn==20.1.0
h5py==2.10.0
HeapDict==1.0.1
heroku==0.1.4
heroku3==4.2.3
hiredis==1.1.0
ibm-db==3.0.3
imbalanced-learn==0.7.0
imblearn==0.0
ipython-sql==0.4.0
jdcal==1.4.1
json5==0.9.5
jupyter==1.0.0
jupyterlab==2.2.6
lightgbm==3.1.1
littleutils==0.2.2
llvmlite==0.35.0
locket==0.2.1
lux==0.8.2
lux-api==0.2.3
lux-widget==0.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.1.0
multidict==5.1.0
multipledispatch==0.6.0
multitasking==0.0.9
mypy-extensions==0.4.3
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.0.3
navigator-updater==0.2.1
numba==0.52.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.7.12
opencensus-context==0.1.2
outdated==0.2.0
pandas-flavor==0.2.0
parso==0.7.0
partd==1.1.0
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pingouin==0.3.9
pkginfo==1.7.0
plotly==4.14.3
pluggy==0.13.1
ply==3.11
protobuf==3.14.0
psycopg2==2.8.6
pulsar==2.0.2
pulsar-odm==0.7.0
py-spy==0.3.4
pyaml==20.4.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3k
pycurl==7.43.0.6
pydeck==0.6.1
PyJWT==2.0.1
pyodbc===4.0.0-unsupported
pyreadline==2.1
pytest==6.2.2
python-engineio==4.0.0
python-sql==1.2.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
ray==1.2.0
redis==3.5.3
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
Rtree==0.9.4
ruamel-yaml==0.15.87
sacremoses==0.0.43
Sastrawi==1.0.1
scikit-image==0.17.2
scikit-optimize==0.8.1
shap==0.38.1
simplegeneric==0.8.1
sip==4.19.13
slicer==0.0.7
smmap==3.0.5
sql==0.4.0
statsmodels==0.12.2
streamlit==0.76.0
tables==3.6.1
tabulate==0.8.7
tensorboardX==2.1
terminado==0.9.2
ticker==0.13
tifffile==2020.10.1
tokenizers==0.9.4
torch==1.7.1
transformers==4.2.1
tune-sklearn==0.2.1
tweepy==3.10.0
twitterscraper==1.6.1
tzlocal==2.1
unicodecsv==0.14.1
validators==0.18.2
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xarray==0.16.2
xgboost==1.3.3
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.6.3
yfinance==0.1.55
zict==2.0.0
zope.event==4.5.0h
